requirements.txt
pandas
numpy
tqdm
scipy
scikit-learn
matplotlib
sentence-transformers
transformers
faiss-cpu
pyserini